backend/orchestrator-service/requirements.txt
uvicorn==0.38.0
fastapi==0.124.2
groq-0.37.1
dotenv==0.9.9
gtts==2.5.4
pillow==11.3.0
moviepy==2.2.1
langchain==1.1.0
langchain-community==0.4.1
langchain-groq==1.1.0
pypdf==6.4.0
sentence-transformers==5.1.2
faiss-cpu==1.13.0
playwright==1.56.0
openai==2.9.0
pyttsx3==2.99
pyaudio==0.2.14
librosa==0.11.0
pydub==0.25.1
pypdf2==3.0.1
langgraph==1.0.4
asyncpg==0.31.0
uvicorn[standard]==3.11
python-multipart==0.0.20
feedparser==6.0.12
opencv-python==4.12.0.88
ultralytics==8.3.252
pandas==2.3.3
google-cloud-storage==3.8.0
openpyxl==3.1.5